Add arm to name and add deadband
