+++ src/main/java/frc/robot/BuildConstants.java
@@ -7,12 +7,12 @@ public final class BuildConstants {
   public static final String MAVEN_GROUP = "";
   public static final String MAVEN_NAME = "robot-2023";
   public static final String VERSION = "unspecified";
-  public static final int GIT_REVISION = 41;
-  public static final String GIT_SHA = "46618c6bfa7b3e733c2e705e036bf715a9717888";
-  public static final String GIT_DATE = "2023-01-19 11:38:05 EST";
-  public static final String GIT_BRANCH = "feature/arm";
-  public static final String BUILD_DATE = "2023-01-20 13:29:29 EST";
-  public static final long BUILD_UNIX_TIME = 1674239369188L;
+  public static final int GIT_REVISION = 85;
+  public static final String GIT_SHA = "11fe5005152d711c1568078607d1d40f9a025ec0";
+  public static final String GIT_DATE = "2023-01-29 14:07:51 EST";
+  public static final String GIT_BRANCH = "feature/non-prototype-arm";
+  public static final String BUILD_DATE = "2023-01-29 14:32:43 EST";
+  public static final long BUILD_UNIX_TIME = 1675020763266L;
   public static final int DIRTY = 1;
 
   private BuildConstants(){}
